+++ Modules/BeaconMinimap.lua
@@ -1,5 +1,5 @@
-local MDT = MDT
 local MDT_NPT = MDT_NPT
+local MDT = MDT_NPT.MDT or MDT
 local pairs, ipairs, tonumber, type = pairs, ipairs, tonumber, type
 local math_max, math_min, math_huge = math.max, math.min, math.huge
 local math_cos, math_sin, math_pi = math.cos, math.sin, math.pi