+++ Modules/Scenario.lua
@@ -1,5 +1,5 @@
-local MDT = MDT
 local MDT_NPT = MDT_NPT
+local MDT = MDT_NPT.MDT or MDT
 ---@type DebugChannel
 local dbg = MDT_NPT.Debug.make("forces", false)